refactor: Split password toggle function
